style(@integration): files reorganization
