+++ packages/@integration/src/http.listener.ts
@@ -1,0 +1,13 @@
+import { httpListener } from '@marblejs/core';
+import { bodyParser$ } from '@marblejs/middleware-body';
+import { loggerDev$, loggerFile$ } from './middlewares/logger.middleware';
+import { api$ } from './effects/api.effects';
+
+export const httpServer = httpListener({
+  middlewares: [
+    loggerDev$,
+    loggerFile$,
+    bodyParser$,
+  ],
+  effects: [api$]
+});
